requirements.txt
Django==1.5.1
Fabric==1.7.0
MySQL-python==1.2.4
PIL==1.1.7
Pillow==2.1.0
South==0.8.1
Unidecode==0.04.14
akismet==0.2.0
argparse==1.2.1
beautifulsoup4==4.3.1
coverage==3.6
django-ckeditor==4.0.2
django-coverage==1.2.4
django-jsonify==0.2.1
django-mptt==0.5.5
django-postman==3.0.2
django-social-auth==0.7.25
django-tagging==0.3.1
django-xmlrpc==0.1.5
ecdsa==0.9
facebook-sdk==0.4.0
factory-boy==2.1.1
fuzzywuzzy==0.2
gunicorn==18.0
httplib2==0.8
ipdb==0.7
ipython==0.13.2
lettuce==0.2.18
meld3==0.6.10
nose==1.3.0
oauth2==1.5.211
paramiko==1.11.0
pycrypto==2.6
pyparsing==1.5.7
python-openid==2.2.5
pytz==2013b
redis==2.8.0
requests==1.2.3
selenium==2.37.2
supervisor==3.0
sure==1.2.2
wsgiref==0.1.2
